v6-n2n-diagnostics/pyproject.toml
[tool.poetry]
name = "n2n_diagnostics"
version = "0.2.0"
description = ""
authors = ["Djura Smits <djura.smits@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
vantage6-client = "^3.3.7"


[tool.poetry.dev-dependencies]
clize = "^4.2.1"
pytest = "^6.2.5"
ruff = "^0.0.244"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Never enforce `E501` (line length violations).
# Ignore star imports
ignore = ["E501", "F403"]